test: gate simulated consumers
